pyproject.toml
[project]
name = "alertmanager-pachka-router"
version = "0.1.0"
description = "Alertmanager webhook -> Pachca messenger router"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.95,<0.100",
  "uvicorn[standard]>=0.30",
  "pydantic>=1.10,<2",
  "pyyaml>=6.0",
  "httpx>=0.27",
]

[tool.uvicorn]
factory = false

